uses: Print cask uses
